requirements.txt
Django==1.3.1
MySQL-python
django-tagging
